requirements.txt
cffi==1.16.0
cryptography==41.0.5
install==1.3.5
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
pycparser==2.21
pydub==0.25.1
